prodsentinel-pipeline/requirements.txt
celery==5.4.0
redis==5.2.1
pyautogen==0.4.0
google-generativeai==0.8.3
sqlalchemy==2.0.45
asyncpg==0.31.0
fastapi==0.127.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
alembic==1.17.2
